sdk-python/pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "agentglass-python"
version = "0.1.0"
description = "Python SDK for AgentGlass local-first telemetry"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "httpx>=0.27.0",
  "pydantic>=2.9.0",
  "typer>=0.12.5"
]

[project.optional-dependencies]
langgraph = ["langgraph>=0.2.0"]
otel = ["opentelemetry-api>=1.20.0", "opentelemetry-sdk>=1.20.0"]

[project.scripts]
agentglass = "agentglass_python.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/agentglass_python"]
